Move `ProjectDiagnosticsEditor::jump` to `Editor::jump`
